feat: update AppDrawerContent to match the design
