+++ app/src/main/java/com/example/karen_and/navigation/AppTopBar.kt
@@ -1,13 +1,20 @@
 package com.example.karen_and.navigation
 
+import android.provider.CalendarContract
+import androidx.compose.foundation.background
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material3.ColorScheme
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarColors
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import com.example.karen_and.R
 
@@ -18,6 +25,7 @@ fun AppTopBar(
     onMenuClick: () -> Unit,
 ) {
     TopAppBar(
+        colors = TopAppBarColors(containerColor = MaterialTheme.colorScheme.primary, actionIconContentColor = Color.White, titleContentColor = Color.White, scrolledContainerColor = Color.Transparent, navigationIconContentColor = Color.White),
         title = {
             Text(text = title)
         },
